src/h_rdt/pyproject.toml
[project]
name = "h-rdt"
version = "0.1.0"
description = "H-RDT training environment"
requires-python = ">=3.11,<3.12"
dependencies = [
  "accelerate==0.34.2",
  "av",
  "deepspeed==0.15.1",
  "datasets==4.1.1",
  "deepdiff",
  "diffusers==0.27.2",
  "draccus",
  "fastapi",
  "h5py==3.11.0",
  "huggingface-hub==0.25.0",
  "imgaug==0.4.0",
  "imageio[ffmpeg]",
  "jsonlines",
  "natsort",
  "numpy==1.26.4",
  "opencv-python==4.10.0.84",
  "packaging==24.0",
  "pandas==2.2.3",
  "pillow",
  "pyarrow",
  "pyyaml",
  "safetensors",
  "scipy",
  "sentencepiece==0.2.0",
  "setuptools>=68",
  "timm==1.0.3",
  "torch==2.4.1",
  "torchvision==0.19.1",
  "tqdm",
  "transformers==4.43.0",
  "uvicorn",
  "wandb==0.21.0",
  "lerobot==0.1.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.uv]
environments = ["sys_platform == 'linux' and python_full_version >= '3.11' and python_full_version < '3.12'"]
package = false

[tool.uv.sources]
lerobot = { git = "https://github.com/huggingface/lerobot.git", rev = "f39652707caed42a7cd5ab36066da5663b9565eb" }

[[tool.uv.dependency-metadata]]
name = "lerobot"
version = "0.1.0"
requires-python = ">=3.10"
requires-dist = []
